Basic Attack Object and Attack Object Tracker
I made a basic structure of how we can represent attack objects, entities on the game board, and how the application might interact with the game board.

I haven't been able to get past an android SDK issue so I haven't been able to test this in android studio.  I actually wrote this in Dr.java and then moved it to android studio.

Still need to add in testing and probably want to move all of this stuff into the same package.