Added the other attack gestures to the game! Also made some really crappy sprites for the other two attacks other than fire.
The fire attack can be invoked by drawing a fire icon but with only two pointy spikes at the top. Basically if you're have difficulty triggering it, try drawing hearts. Fast hearts.

The water attack can be invoked by drawing a very wide 'M' or 'W', Like a sin wave, but with 1.5 periods.

The ground attack can be invoked by drawing a mound of dirt. Just draw a hump. That's it LMAO