Anti-Attack Spam
Added a limit of maxObjects to each team in a game. Note that I didn't change the calls in GameBoard itself, so this effectively doubles the max count of objects (I believe we now have a max of 10 objects per side).

This stops us from starving the AI of attacks (or vice versa), but it doesn't necessarily stop ground attack spam from being a viable tactic.